import streamlit as st.py
import streamlit as st
import pandas as pd
import numpy as np

class NSSFrameworkAgent:
    def __init__(self):
        # Load NSS framework data and assessment criteria
        self.nss_data = self.load_nss_framework()
        self.assessment_questions = self.load_assessment_questions()
    
    def load_nss_framework(self):
        # Load comprehensive NSS framework details
        pass
    
    def load_assessment_questions(self):
        # Load personality/organizational assessment questions
        questions = [
            # Categorized questions mapping to NSS dimensions
            {
                "category": "Leadership Style",
                "questions": [
                    "How do you typically make decisions?",
                    "What's your approach to problem-solving?"
                ]
            },
            # More categories...
        ]
        return questions
    
    def conduct_nss_assessment(self):
        # Interactive assessment flow
        st.title("NSS Framework Personality Assessment")
        
        # Dynamic questionnaire generation
        assessment_results = {}
        for category in self.assessment_questions:
            st.subheader(category['category'])
            for question in category['questions']:
                response = st.radio(question, 
                    options=[
                        "Strongly Disagree", 
                        "Disagree", 
                        "Neutral", 
                        "Agree", 
                        "Strongly Agree"
                    ])
                # Score and categorize responses
                assessment_results[question] = response
        
        return assessment_results
    
    def map_nss_stage(self, assessment_results):
        # Algorithmic mapping of results to NSS stage
        # Complex scoring mechanism
        # For simplicity, let's assume it returns a stage based on some logic
        return "Nail It"  # Example stage
    
    def generate_recommendations(self, mapped_stage):
        # Generate personalized recommendations
        recommendations = {
            "Nail It": {
                "Organization": ["Focus on core competencies", "Streamline operations"],
                "Leader": ["Develop leadership skills", "Enhance decision-making"],
                "Team": ["Build a strong team culture", "Encourage collaboration"]
            },
            "Scale It": {
                "Organization": ["Expand market reach", "Invest in technology"],
                "Leader": ["Strategic planning", "Visionary leadership"],
                "Team": ["Upskill team members", "Foster innovation"]
            },
            "Sail It": {
                "Organization": ["Maintain stability", "Optimize processes"],
                "Leader": ["Sustain leadership", "Mentor successors"],
                "Team": ["Ensure team satisfaction", "Promote work-life balance"]
            }
        }
        return recommendations.get(mapped_stage, {})
    
    def visualization_dashboard(self, assessment_results, stage):
        # Create interactive visualizations of assessment
        pass

def main():
    nss_agent = NSSFrameworkAgent()
    
    st.sidebar.title("NSS Framework Journey")
    
    # Workflow steps
    assessment_results = nss_agent.conduct_nss_assessment()
    mapped_stage = nss_agent.map_nss_stage(assessment_results)
    recommendations = nss_agent.generate_recommendations(mapped_stage)
    
    # Display results and recommendations
    st.success(f"Your Organization is in the {mapped_stage} Stage")
    st.write(recommendations)

if __name__ == "__main__":
    main()