requirements-file.txt
# Core dependencies
numpy>=1.20.0
pyyaml>=6.0
requests>=2.28.0

# Embeddings and vector storage
scikit-learn>=1.0.0
sentence-transformers>=2.2.0

# LLM API clients
anthropic>=0.5.0
openai>=1.0.0

# Data processing and visualization
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.12.0

# Web tools
beautifulsoup4>=4.10.0
lxml>=4.6.0

# Development and testing
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
isort>=5.10.0
pylint>=2.12.0
mypy>=0.950
